Decompose TooManyFunctions: split VMs, use cases, and delegates
- Split WorkflowTemplateUseCases.kt into 7 individual files (#736)
- Extract GenerationExecutionDelegate from ComfyUIGenerationViewModel (#736)
- Extract SearchFilterDelegate from ModelSearchViewModel (#736)
- Extract DetailNotesTagsDelegate from ModelDetailViewModel (#736)
- Extract GalleryGenerationDelegate from ExternalServerGalleryViewModel (#736)
- Extract DatasetExportDelegate from DatasetDetailViewModel (#736)
- Remove @Suppress("TooManyFunctions") from 4 VMs (Search, ComfyUIGeneration, WorkflowTemplate, ComfyHubBrowser/Detail)

Closes #736
